restructure Shop,Category Modules,removed BrandProduct Module
