--- Modules/BrandProduct/Database/Seeders/BrandProductDatabaseSeeder.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Modules\BrandProduct\Database\Seeders;
-
-use Illuminate\Database\Seeder;
-use Illuminate\Database\Eloquent\Model;
-
-class BrandProductDatabaseSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        Model::unguard();
-
-        // $this->call("OthersTableSeeder");
-    }
-}
