--- Modules/BrandProduct/Resources/views/index.blade.php
@@ -1,9 +1,0 @@
-@extends('brandproduct::layouts.master')
-
-@section('content')
-    <h1>Hello World</h1>
-
-    <p>
-        This view is loaded from module: {!! config('brandproduct.name') !!}
-    </p>
-@endsection
