--- Modules/BrandProduct/Routes/web.php
@@ -1,16 +1,0 @@
-<?php
-
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
-Route::prefix('brandproduct')->group(function() {
-    Route::get('/', 'BrandProductController@index');
-});
