--- Modules/BrandProduct/module.json
@@ -1,11 +1,0 @@
-{
-    "name": "BrandProduct",
-    "alias": "brandproduct",
-    "description": "",
-    "keywords": [],
-    "priority": 0,
-    "providers": [
-        "Modules\\BrandProduct\\Providers\\BrandProductServiceProvider"
-    ],
-    "files": []
-}
