--- Modules/Category/Database/Seeders/CategoryDatabaseSeeder.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Modules\Category\Database\Seeders;
-
-use Illuminate\Database\Seeder;
-use Illuminate\Database\Eloquent\Model;
-
-class CategoryDatabaseSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        Model::unguard();
-
-        // $this->call("OthersTableSeeder");
-    }
-}
