--- Modules/Category/Resources/views/index.blade.php
@@ -1,9 +1,0 @@
-@extends('category::layouts.master')
-
-@section('content')
-    <h1>Hello World</h1>
-
-    <p>
-        This view is loaded from module: {!! config('category.name') !!}
-    </p>
-@endsection
