--- Modules/Shop/Database/Seeders/ShopDatabaseSeeder.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Modules\Shop\Database\Seeders;
-
-use Illuminate\Database\Seeder;
-use Illuminate\Database\Eloquent\Model;
-
-class ShopDatabaseSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        Model::unguard();
-
-        // $this->call("OthersTableSeeder");
-    }
-}
