--- Modules/Shop/Resources/views/index.blade.php
@@ -1,9 +1,0 @@
-@extends('shop::layouts.master')
-
-@section('content')
-    <h1>Hello World</h1>
-
-    <p>
-        This view is loaded from module: {!! config('shop.name') !!}
-    </p>
-@endsection
