--- Modules/Shop/package.json
@@ -1,16 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build"
-  },
-  "devDependencies": {
-    "axios": "^0.21.4",
-    "dotenv": "^10.0.0",
-    "dotenv-expand": "^5.1.0",
-    "laravel-vite-plugin": "^0.6.0",
-    "lodash": "^4.17.21",
-    "postcss": "^8.3.7",
-    "vite": "^3.0.9"
-  }
-}
